model_handler_service/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model_handler_service"
version = "0.1.0"
description = "A service for processing clothing images and determining body types using machine learning models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Chi Bepoosham!", email = "xxxxx@chibepoosham.com"},
]
dependencies = [
    "redis==5.2.1",
    "datetime==5.5",
    "tensorflow",
    "torch",
    "torchvision",
    "numpy",
    "ultralytics",
    "opencv-python",
    "pillow",
    "requests",
    "python-dotenv>=1.0.0",
    "scikit-learn",
    "mediapipe",
    "flask"
]
classifiers = [
    'Development Status :: 3 - Alpha',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Topic :: Scientific/Engineering :: Image Processing',
]

[project.urls]
Homepage = "https://github.com/chi-bepoosham/model_handler_service"
Repository = "https://github.com/chi-bepoosham/model_handler_service"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"] 